requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.6
distlib==0.3.8
Django==3.0.8
django_vite_plugin==3.0.0
djangorestframework==3.11.0
filelock==3.13.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==10.1.0
platformdirs==4.1.0
pylint==2.5.3
python-dotenv==1.0.0
pytz==2020.1
setuptools==69.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tzdata==2023.3
virtualenv==20.25.0
wrapt==1.12.1
